As per WP.org requirements updated the following:
One or more tags were ignored `screencast`. Please limit your plugin to 5 tags.
The License field is missing. A GPLv2 or later compatible license should be specified.
The Short Description section is too long and was truncated. A maximum of 150 characters is supported.

Other than that:
the "Tested up to" was updated to 6.7.1
the minimum PHP requirement updated to 7.4